RuleSet objects are now configurable.
All rulecondition functionality is contained inside of the base RuleCondition class. NeighborState and CurrentState are deprecated and are present to allow GOL cells to work.

RuleSets can have rules and ruleconditions defined in them. The targets field (which neighbors this rule affects) is currently set using an array of indices. This should be changed eventually to allow for easier modifications.